Added authorization in batches,courses,projects,students,users,Updated App.js made StudentForm and UserForm render in studentTable and userList respectively, updated Login,Added back button in Both student and userforms. Added Add button in student and user tables

+++ backend/Routes/courses.js
@@ -1,9 +1,10 @@
 const router = require("express").Router();
 
 const courseData = require("../Models/course");
+const auth = require("../middleware/Auth");
 
 // get course
-router.get("/course", async (req, res) => {
+router.get("/course", auth, async (req, res) => {
   try {
     let courses = await courseData.find();
     res.json(courses);
@@ -13,12 +14,16 @@ router.get("/course", async (req, res) => {
 });
 
 //add new course
-router.post("/course", async (req, res) => {
+router.post("/course", auth, async (req, res) => {
   try {
-    console.log(req.body);
-    const newCourse = courseData(req.body);
-    await newCourse.save();
-    res.json({ message: "Course added successfully" });
+    if (req.body.role === "Admin") {
+      console.log(req.body);
+      const newCourse = courseData(req.body);
+      await newCourse.save();
+      res.json({ message: "Course added successfully" });
+    } else {
+      res.json({ message: "Access denied" });
+    }
   } catch (error) {
     res.json({ message: "Unable to post" });
   }
@@ -26,11 +31,15 @@ router.post("/course", async (req, res) => {
 
 // Delete course
 
-router.delete("/course/:id", async (req, res) => {
+router.delete("/course/:id", auth, async (req, res) => {
   try {
-    const { id } = req.params;
-    await courseData.findByIdAndDelete(id);
-    res.json({ message: "Course deleted successfully" });
+    if (req.body.role === "Admin") {
+      const { id } = req.params;
+      await courseData.findByIdAndDelete(id);
+      res.json({ message: "Course deleted successfully" });
+    } else {
+      res.json({ message: "Access denied" });
+    }
   } catch (error) {
     res.json({ message: "unable to delete", err: error.message });
   }
